create functions for create day and working graphics

+++ app/crud/working_graphic.py
@@ -1,8 +1,23 @@
-from fastapi import HTTPException
-from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.future import select
+from sqlalchemy.orm import selectinload
+from sqlalchemy.ext.asyncio import AsyncSession
+
 from ..models.working_graphic import WorkingGraphic, Day
 from ..schemas.working_graphic import WorkingGraphicCreate, WorkingGraphicUpdate, DayCreate, DayUpdate
+
+
+async def create_day(db: AsyncSession, day: DayCreate, working_graphic_id: int):
+    try:
+        db_day = Day(**day.model_dump(), working_graphic_id=working_graphic_id)
+        db.add(db_day)
+        await db.commit()
+        await db.refresh(db_day)
+
+    except Exception as e:
+        await db.rollback()
+        raise e
+
+    return db_day
 
 
 async def create_working_graphic(db: AsyncSession, working_graphic: WorkingGraphicCreate):
@@ -12,17 +27,16 @@ async def create_working_graphic(db: AsyncSession, working_graphic: WorkingGraph
         await db.commit()
         await db.refresh(db_working_graphic)
 
-        for day in working_graphic.days:
-            db_day = Day(**day.model_dump(), working_graphic_id=db_working_graphic.id)
-            db.add(db_day)
-
-        await db.commit()
-
-        return db_working_graphic
     except Exception as e:
         await db.rollback()
-        raise HTTPException(status_code=500, detail=str(e))
+        raise e
 
+    return db_working_graphic
+
+
+async def get_days(db: AsyncSession, working_graphic_id: int):
+    result = await db.execute(select(Day).filter_by(working_graphic_id=working_graphic_id))
+    return result.scalars().all()
 
 async def get_working_graphics(db: AsyncSession, skip: int = 0, limit: int = 10):
     result = await db.execute(select(WorkingGraphic).offset(skip).limit(limit))
@@ -33,7 +47,7 @@ async def get_working_graphic(db: AsyncSession, working_graphic_id: int):
     result = await db.execute(select(WorkingGraphic).filter_by(id=working_graphic_id))
     working_graphic = result.scalar_one_or_none()
     if not working_graphic:
-        raise HTTPException(status_code=404, detail="Working graphic not found")
+        raise Exception("Working graphic not found")
     return working_graphic
 
 
@@ -48,6 +62,8 @@ async def update_working_graphic(db: AsyncSession, working_graphic_id: int, work
 
 async def delete_working_graphic(db: AsyncSession, working_graphic_id: int):
     db_working_graphic = await get_working_graphic(db, working_graphic_id)
+    if not db_working_graphic:
+        raise Exception("Working graphic not found")
     await db.delete(db_working_graphic)
     await db.commit()
     return {"message": f"Working graphic {working_graphic_id} deleted"}
